requirements.txt
requests==2.31.0
python-telegram-bot==22.6
flask==2.3.2
